refactor(services): migrate simple shim callsites to canonical paths (#228)
Bulk-migrates `from app.services.<shim>` imports to the new canonical
hexagonal locations for the 12 pure-re-export shims (no behavioural
change — the shims themselves are removed in the follow-up commit):

- authorization_service     -> app.servers.application.authorization
- minecraft_server          -> app.servers.application.minecraft_server
- simplified_sync           -> app.servers.application.simplified_sync
- real_time_server_commands -> app.servers.application.real_time_server_commands
- server_properties_generator -> app.servers.application.server_properties_generator
- jar_cache_manager         -> app.versions.application.jar_cache_manager
- java_compatibility        -> app.versions.application.java_compatibility
- minecraft_api_service     -> app.versions.application.minecraft_api_service
- version_manager           -> app.versions.application.version_manager
- encoding_handler          -> app.files.application.encoding_handler
- visibility_service        -> app.core.visibility.application.service
- visibility_migration_service -> app.core.visibility.application.migration

Plus `ServerService` import in tests/unit/services/test_server_service.py.
