+++ tests/integration/test_daemon_lifecycle_comprehensive.py
@@ -12,7 +12,7 @@ from unittest.mock import AsyncMock, Mock, patch
 import pytest
 
 from app.servers.models import Server, ServerStatus, ServerType
-from app.services.minecraft_server import MinecraftServerManager, ServerProcess
+from app.servers.application.minecraft_server import MinecraftServerManager, ServerProcess
 
 pytestmark = pytest.mark.slow
