+++ tests/integration/test_minecraft_server_monitoring.py
@@ -12,7 +12,7 @@ import pytest
 import pytest_asyncio
 
 from app.servers.models import ServerStatus
-from app.services.minecraft_server import MinecraftServerManager, ServerProcess
+from app.servers.application.minecraft_server import MinecraftServerManager, ServerProcess
 
 pytestmark = pytest.mark.slow
