+++ tests/integration/test_process_persistence.py
@@ -21,7 +21,7 @@ import psutil
 import pytest
 
 from app.servers.models import ServerStatus
-from app.services.minecraft_server import MinecraftServerManager, ServerProcess
+from app.servers.application.minecraft_server import MinecraftServerManager, ServerProcess
 
 pytestmark = [pytest.mark.asyncio, pytest.mark.slow]
