+++ tests/unit/servers/routers/test_import_export.py
@@ -95,7 +95,7 @@ class TestImportExportRouter:
 
     def test_import_server_regular_user_authorization_check(self, test_user):
         """Test that regular users pass authorization check for server import (Phase 1: shared resource model)"""
-        from app.services.authorization_service import AuthorizationService
+        from app.servers.application.authorization import AuthorizationService
 
         # Phase 1: Regular users should be able to create (import) servers
         assert AuthorizationService.can_create_server(test_user) is True, (
