+++ tests/unit/servers/test_simplified_sync.py
@@ -12,7 +12,7 @@ from app.core.database import Base
 from app.servers.models import Server, ServerStatus, ServerType
 from app.servers.schemas import ServerUpdateRequest
 from app.servers.service import server_service
-from app.services.simplified_sync import simplified_sync_service
+from app.servers.application.simplified_sync import simplified_sync_service
 from app.users.domain.value_objects import Role
 from app.users.models import User
