+++ tests/unit/services/test_minecraft_server.py
@@ -11,7 +11,7 @@ from unittest.mock import AsyncMock, Mock, patch
 import pytest
 
 from app.servers.models import Server, ServerStatus, ServerType
-from app.services.minecraft_server import MinecraftServerManager, ServerProcess
+from app.servers.application.minecraft_server import MinecraftServerManager, ServerProcess
 
 
 class TestMinecraftServerManagerSimpleCoverage:
