+++ tests/unit/services/test_phase1_compliance.py
@@ -16,7 +16,7 @@ from fastapi import HTTPException
 from sqlalchemy.orm import Session
 
 from app.servers.models import Server, ServerStatus, ServerType
-from app.services.authorization_service import AuthorizationService
+from app.servers.application.authorization import AuthorizationService
 from app.users.domain.value_objects import Role
 from app.users.models import User
