+++ tests/unit/services/test_real_time_server_commands.py
@@ -7,7 +7,7 @@ import pytest
 from app.core.security import SecurityError
 from app.groups.models import GroupType
 from app.servers.models import ServerStatus
-from app.services.real_time_server_commands import (
+from app.servers.application.real_time_server_commands import (
     RealTimeServerCommandService,
     real_time_server_commands,
 )
